Update PasswordModal Refresh Function
